finish call handling unit tests
